Add support for status attribute when using finding_count
This allows users to count findings based on threatLevel, as well as based on one or more discrete status values.

Example in XML:
   <finding_count threatLevel="High" status="new unresolved"/>

This will count the number of findings having attribute threatLevel="High", AND a status of either "new" OR "unresolved".

This change is backwards-compatible: Using finding_count without the status attribute will show the same results as before.